requirements.txt
langchain>=0.2.0,<0.3.0
langchain-google-genai>=1.0.5,<2.0.0
langchain-community>=0.2.0,<0.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.1,<2.0.0
